Completed testcases and rubocop offence correction.
