+++ app/controllers/hotels_controller.rb
@@ -9,8 +9,7 @@ class HotelsController < ApplicationController
   def create
     check
     begin
-      Hotel.create(create_attrs(params))
-      @hotel = Hotel.last
+      @hotel = Hotel.create(create_attrs(params))
     rescue StandardError => error
       flash_error(error, 'new')
     else
@@ -62,8 +61,8 @@ class HotelsController < ApplicationController
   end
 
   def image_room_create(params)
-    Image.create(image_attr(params))
-    @hotel.update_attributes!(image_id: Image.last.id)
+    @image = Image.create(image_attr(params))
+    @hotel.update_attributes!(image_id: @image.id)
     iteration = 101..(params[:hotels][:no_of_rooms].to_i + 100)
     iteration.each do |room_no|
       Room.create(room_attrs(room_no))
@@ -74,7 +73,7 @@ class HotelsController < ApplicationController
   def image_attr(params)
     {
       image: params[:hotels][:imageable],
-      imageable_id: Hotel.last.id,
+      imageable_id: @hotel.id,
       imageable_type: 'Hotel'
     }
   end
@@ -94,7 +93,7 @@ class HotelsController < ApplicationController
 
   def room_attrs(room_no)
     { no: room_no,
-      hotel_id: Hotel.last.id,
+      hotel_id: @hotel.id,
       estimated_time: '01:00:00',
       status: 'dirty' }
   end
