+++ app/views/login/new.html.erb
@@ -1,10 +1,12 @@
 <body style="text-align:center">
 <br><br>
 <h1>&nbsp;Login</h1><br>
+<div class="alert alert-info">
 <%= form_tag(:controller => "login", :action => "create") do %>
 	Username : <%= text_field(:users, :name) %><br /><br />
 	Password  &nbsp;: <%= password_field(:users, :password)%><br /><br/>
 	<p><%= submit_tag "Submit",:class => 'btn btn-large btn-primary'%></p>
-<% end %><br>
+<% end %></div>
+<br>
 <%= link_to "Sign in with Google", "/auth/google_oauth2", id: "sign_in" %>
 </body>
