+++ app/views/users/_edit_email.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag({:controller => "users", :action => "update", :id=>user.id}, remote: true) do%>
+<%= form_tag({:controller => "users", :action => "update", :id=>@user.id}, remote: true) do%>
 	<%= email_field(:users, :email, value: @user.email)%>
 	<%= submit_tag "Update"%>
 <%end%>
