+++ app/views/users/_edit_image.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag({ :controller => "users", :action => "update", :id=> user.id }, { :remote => true , :multipart => true }) do%>
+<%= form_tag({ :controller => "users", :action => "update", :id=> @user.id }, { :remote => true , :multipart => true }) do%>
 	<%= file_field(:users, :imageable)%>
 	<%= submit_tag "Update"%>
 <%end%>
