+++ app/views/users/_edit_name.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag({:controller => "users", :action => "update", :id=>user.id}, remote: true) do%>
+<%= form_tag({:controller => "users", :action => "update", :id=>@user.id}, remote: true) do%>
 	<%= text_field(:users, :name, value: @user.name)%>
 	<%= submit_tag "Update"%>
 <%end%>
