+++ spec/app/models/ability_spec.rb
@@ -1,5 +1,6 @@
 require 'spec_helper'
 require 'rails_helper'
+require 'cancan/matchers'
 
 describe 'ability' do
   let(:subject) { Ability.new(user) }
